# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "电池进销存管理系统的后端API接口文档",
        "title": "电池进销存管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8036",
    "basePath": "/jxc/v1",
    "paths": {
...
        "/inbound/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取入库订单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "入库管理"
                ],
                "summary": "获取所有入库订单",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的入库订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "入库管理"
                ],
                "summary": "创建入库订单",
                "parameters": [
                    {
                        "description": "入库订单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateInboundOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inbound/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据订单ID获取入库订单详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "入库管理"
                ],
                "summary": "根据ID获取入库订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID更新入库订单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "入库管理"
                ],
                "summary": "更新入库订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "入库订单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InboundOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID删除入库订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "入库管理"
                ],
                "summary": "删除入库订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inventory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有电池分类的库存信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "获取所有库存",
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inventory/{categoryId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据电池分类ID获取该分类的库存信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "库存管理"
                ],
                "summary": "根据分类ID获取库存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "categoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateInboundOrderItem": {
            "type": "object",
            "required": [
                "category_id",
                "gross_weight",
                "tare_weight",
                "unit_price"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "gross_weight": {
                    "type": "number"
                },
                "tare_weight": {
                    "type": "number",
                    "minimum": 0
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
        "models.CreateInboundOrderRequest": {
            "type": "object",
            "required": [
                "items",
                "supplier_name"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateInboundOrderItem"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "supplier_name": {
                    "type": "string"
                }
            }
        },
        "models.InboundOrder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "created_by": {
                    "description": "创建人",
                    "type": "integer"
                },
                "id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "notes": {
                    "description": "备注",
                    "type": "string"
                },
                "order_no": {
                    "description": "订单号",
                    "type": "string"
                },
                "status": {
                    "description": "'completed', 'cancelled'",
                    "type": "string"
                },
                "supplier_name": {
                    "description": "供应商名称",
                    "type": "string"
                },
                "total_amount": {
                    "description": "总金额",
                    "type": "number"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}